#31 Refactor database and related code to use Spring JPA

+++ Server/src/main/java/edu/sdccd/cisc191/f/server/AccountRepository.java
@@ -1,6 +1,5 @@
 package edu.sdccd.cisc191.f.server;
 
-import edu.sdccd.cisc191.f.Account;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -11,6 +10,6 @@ public interface AccountRepository extends CrudRepository<Account, Long> {
 
     Account findByID(long id);
     Account findAccountByCardNumber(long cardNumber);
-    List<Account> findAccountsByBalanceEquals(int balance);
-    List<Account> findAccountsByBalanceGreaterThan(int bigMoney);
+    List<Account> findAccountsByBalanceEquals(double balance);
+    List<Account> findAccountsByBalanceGreaterThan(double bigMoney);
 }
